requirements.txt
pytest~=8.3.5
allure-pytest==2.14.0
selene>=2.0.0rc9
allure-python-commons~=2.14.0